+++ src/pygeoprocessing/__init__.py
@@ -11,8 +11,10 @@ import pkg_resources.extern.packaging.version
 from . import geoprocessing
 from .geoprocessing import _assert_is_valid_pixel_size
 from .geoprocessing import align_and_resize_raster_stack
+from .geoprocessing import build_overviews
 from .geoprocessing import calculate_disjoint_polygon_set
 from .geoprocessing import convolve_2d
+from .geoprocessing import create_raster_from_bounding_box
 from .geoprocessing import create_raster_from_vector_extents
 from .geoprocessing import distance_transform_edt
 from .geoprocessing import get_gis_type
@@ -25,6 +27,7 @@ from .geoprocessing import merge_bounding_box_list
 from .geoprocessing import new_raster_from_base
 from .geoprocessing import numpy_array_to_raster
 from .geoprocessing import raster_calculator
+from .geoprocessing import raster_reduce
 from .geoprocessing import raster_to_numpy_array
 from .geoprocessing import rasterize
 from .geoprocessing import ReclassificationMissingValuesError