+++ src/pygeoprocessing/geoprocessing.py
@@ -25,6 +25,7 @@ import shapely.ops
 import shapely.prepared
 import shapely.wkb
 from osgeo import gdal
+from osgeo import gdalconst
 from osgeo import ogr
 from osgeo import osr
 
@@ -79,6 +80,23 @@ _GDAL_TYPE_TO_NUMPY_LOOKUP = {
     gdal.GDT_CFloat32: numpy.csingle,
     gdal.GDT_CFloat64: numpy.complex64,
 }
+
+# GDAL's python API recognizes certain strings but the only way to retrieve
+# those strings is to do this conversion of gdalconst.GRA_* types to the
+# human-readable labels via gdal.WarpOptions like so.
+_GDAL_WARP_ALGORITHMS = []
+for _warp_algo in (_attrname for _attrname in dir(gdalconst)
+                   if _attrname.startswith('GRA_')):
+    # Appends ['-r', 'near'] to _GDAL_WARP_ALGORITHMS if _warp_algo is
+    # gdalconst.GRA_NearestNeighbor.  See gdal.WarpOptions for the dict
+    # defining this mapping.
+    gdal.WarpOptions(options=_GDAL_WARP_ALGORITHMS,
+                     resampleAlg=getattr(gdalconst, _warp_algo))
+_GDAL_WARP_ALGORITHMS = set(_GDAL_WARP_ALGORITHMS)
+_GDAL_WARP_ALGORITHMS.discard('-r')
+_GDAL_WARP_ALGOS_FOR_HUMAN_EYES = "|".join(_GDAL_WARP_ALGORITHMS)
+LOGGER.debug(
+    f'Detected warp algorithms: {", ".join(_GDAL_WARP_ALGOS_FOR_HUMAN_EYES)}')
 
 
 def raster_calculator(
@@ -475,7 +493,8 @@ def raster_calculator(
             pixels_processed += blocksize[0] * blocksize[1]
             last_time = _invoke_timed_callback(
                 last_time, lambda: LOGGER.info(
-                    '%.1f%% complete',
+                    '%s %.1f%% complete',
+                    os.path.basename(target_raster_path),
                     float(pixels_processed) / n_pixels * 100.0),
                 _LOGGING_PERIOD)
 
@@ -532,6 +551,107 @@ def raster_calculator(
                 raise exception
             except queue.Empty:
                 pass
+
+
+def array_equals_nodata(array, nodata):
+    """Check for the presence of ``nodata`` values in ``array``.
+
+    The comparison supports ``numpy.nan`` and unset (``None``) nodata values.
+
+    Args:
+        array (numpy array): the array to mask for nodata values.
+        nodata (number): the nodata value to check for. Supports ``numpy.nan``.
+
+    Returns:
+        A boolean numpy array with values of 1 where ``array`` is equal to
+        ``nodata`` and 0 otherwise.
+    """
+    # If nodata is undefined, nothing matches nodata.
+    if nodata is None:
+        return numpy.zeros(array.shape, dtype=bool)
+
+    # comparing an integer array against numpy.nan works correctly and is
+    # faster than using numpy.isclose().
+    if numpy.issubdtype(array.dtype, numpy.integer):
+        return array == nodata
+    return numpy.isclose(array, nodata, equal_nan=True)
+
+
+def raster_reduce(function, raster_path_band, initializer, mask_nodata=True,
+                  largest_block=_LARGEST_ITERBLOCK):
+    """Cumulatively apply a reducing function to each block of a raster.
+
+    This effectively reduces the entire raster to a single value, but it works
+    by blocks to be memory-efficient.
+
+    The ``function`` signature should be ``function(aggregator, block)``, where
+    ``aggregator`` is the aggregated value so far, and ``block`` is a flattened
+    numpy array containing the data from the block to reduce next.
+
+    ``function`` is called once on each block. On the first ``function`` call,
+    ``aggregator`` is initialized with ``initializer``. The return value from
+    each ``function`` call is passed in as the ``aggregator`` argument to the
+    subsequent ``function`` call. When all blocks have been reduced, the return
+    value of the final ``function`` call is returned.
+
+    Example:
+        Calculate the sum of all values in a raster::
+
+            raster_reduce(lambda total, block: total + numpy.sum(block),
+                          (raster_path, 1), 0)
+
+        Calculate a histogram of all values in a raster::
+
+            def add_to_histogram(histogram, block):
+                return histogram + numpy.histogram(block, bins=10)[0]
+
+            raster_reduce(add_to_histogram, (raster_path, 1), numpy.zeros(10))
+
+        Calculate the sum of all values in a raster, excluding nodata::
+
+            nodata = pygeoprocessing.get_raster_info(raster_path)['nodata'][0]
+            def sum_excluding_nodata(total, block):
+                return total + numpy.sum(block[block != nodata])
+
+            raster_reduce(sum_excluding_nodata, (raster_path, 1), 0)
+
+    Args:
+        function (func): function to apply to each raster block
+        raster_path_band (tuple): (path, band) tuple of the raster to reduce
+        initializer (obj): value to initialize the aggregator for the
+            first function call
+        mask_nodata (bool): if True, mask out nodata before aggregating. A
+            flattened array of non-nodata pixels from each block is passed to
+            the ``function``. if False, each block is passed to the
+            ``function`` without masking.
+        largest_block (int): largest block parameter to pass to ``iterblocks``
+
+    Returns:
+        aggregate value, the final value returned from ``function``
+    """
+    aggregator = initializer
+    last_time = time.time()
+    pixels_processed = 0
+    raster_info = get_raster_info(raster_path_band[0])
+    x_size, y_size = raster_info['raster_size']
+    n_pixels = x_size * y_size
+    for (_, block) in iterblocks(raster_path_band,
+                                 largest_block=largest_block):
+        if mask_nodata:
+            data = block[~array_equals_nodata(
+                block, raster_info['nodata'][raster_path_band[1] - 1])]
+        else:
+            data = block.flatten()
+        aggregator = function(aggregator, data)
+        pixels_processed += block.size
+        last_time = _invoke_timed_callback(
+            last_time, lambda: LOGGER.info(
+                f'{raster_path_band[0]} reduce '
+                f'{pixels_processed / n_pixels * 100:.1f}%% complete'),
+            _LOGGING_PERIOD)
+
+    LOGGER.info('100.0%% complete')
+    return aggregator
 
 
 def align_and_resize_raster_stack(
@@ -1028,52 +1148,98 @@ def create_raster_from_vector_extents(
                 LOGGER.warning(error)
         layer = None
 
+    target_srs_wkt = vector.GetLayer(0).GetSpatialRef().ExportToWkt()
+    vector = None
+
     if shp_extent is None:
         raise ValueError(
             f'the vector at {base_vector_path} has no geometry, cannot '
             f'create a raster from these extents')
 
-    # round up on the rows and cols so that the target raster encloses the
-    # base vector
-    n_cols = int(numpy.ceil(
-        abs((shp_extent[1] - shp_extent[0]) / target_pixel_size[0])))
-    n_cols = max(1, n_cols)
+    create_raster_from_bounding_box(
+        target_bounding_box=[
+            shp_extent[0], shp_extent[2], shp_extent[1], shp_extent[3]
+        ],
+        target_raster_path=target_raster_path,
+        target_pixel_size=target_pixel_size,
+        target_pixel_type=target_pixel_type,
+        target_srs_wkt=target_srs_wkt,
+        target_nodata=target_nodata,
+        fill_value=fill_value,
+        raster_driver_creation_tuple=raster_driver_creation_tuple
+    )
 
-    n_rows = int(numpy.ceil(
-        abs((shp_extent[3] - shp_extent[2]) / target_pixel_size[1])))
-    n_rows = max(1, n_rows)
+
+def create_raster_from_bounding_box(
+        target_bounding_box, target_raster_path, target_pixel_size,
+        target_pixel_type, target_srs_wkt, target_nodata, fill_value=None,
+        raster_driver_creation_tuple=DEFAULT_GTIFF_CREATION_TUPLE_OPTIONS):
+    """Create a raster from a given bounding box.
+
+    Args:
+        target_bounding_box (tuple): a 4-element iterable of (minx, miny,
+            maxx, maxy) in projected units matching the SRS of
+            ``target_srs_wkt``.
+        target_raster_path (string): The path to where the new raster should be
+            created on disk.
+        target_pixel_size (tuple): A 2-element tuple of the (x, y) pixel size
+            of the target raster.  Elements are in units of the target SRS.
+        target_pixel_type (int): The GDAL GDT_* type of the target raster.
+        target_srs_wkt (string): The SRS of the target raster, in Well-Known
+            Text format.
+        target_nodata (float): The nodata value of the target raster, or
+            ``None`` if no nodata value is to be set.
+        fill_value=None (number): If provided, the value that the target raster
+            should be filled with.
+        raster_driver_creation_tuple (tuple): a tuple containing a GDAL driver
+            name string as the first element and a GDAL creation options
+            tuple/list as the second. Defaults to a GTiff driver tuple
+            defined at geoprocessing.DEFAULT_GTIFF_CREATION_TUPLE_OPTIONS.
+
+    Returns:
+        ``None``
+    """
+    if target_pixel_type not in _VALID_GDAL_TYPES:
+        raise ValueError(
+            f'Invalid target type, should be a gdal.GDT_* type, received '
+            f'"{target_pixel_type}"')
+
+    bbox_minx, bbox_miny, bbox_maxx, bbox_maxy = target_bounding_box
 
     driver = gdal.GetDriverByName(raster_driver_creation_tuple[0])
     n_bands = 1
+    n_cols = int(numpy.ceil(
+        abs((bbox_maxx - bbox_minx) / target_pixel_size[0])))
+    n_cols = max(1, n_cols)
+
+    n_rows = int(numpy.ceil(
+        abs((bbox_maxy - bbox_miny) / target_pixel_size[1])))
+    n_rows = max(1, n_rows)
+
     raster = driver.Create(
         target_raster_path, n_cols, n_rows, n_bands, target_pixel_type,
         options=raster_driver_creation_tuple[1])
-    raster.GetRasterBand(1).SetNoDataValue(target_nodata)
+    raster.SetProjection(target_srs_wkt)
 
     # Set the transform based on the upper left corner and given pixel
-    # dimensions
-    if target_pixel_size[0] < 0:
-        x_source = shp_extent[1]
-    else:
-        x_source = shp_extent[0]
-    if target_pixel_size[1] < 0:
-        y_source = shp_extent[3]
-    else:
-        y_source = shp_extent[2]
+    # dimensions.
+    x_source = bbox_maxx if target_pixel_size[0] < 0 else bbox_minx
+    y_source = bbox_maxy if target_pixel_size[1] < 0 else bbox_miny
     raster_transform = [
-        x_source, target_pixel_size[0], 0.0,
-        y_source, 0.0, target_pixel_size[1]]
+        x_source, target_pixel_size[0], 0,
+        y_source, 0, target_pixel_size[1]]
     raster.SetGeoTransform(raster_transform)
 
-    # Use the same projection on the raster as the shapefile
-    raster.SetProjection(vector.GetLayer(0).GetSpatialRef().ExportToWkt())
-
-    # Initialize everything to nodata
+    # Fill the band if requested.
+    band = raster.GetRasterBand(1)
     if fill_value is not None:
-        band = raster.GetRasterBand(1)
         band.Fill(fill_value)
-        band = None
-    vector = None
+
+    # Set the nodata value.
+    if target_nodata is not None:
+        band.SetNoDataValue(float(target_nodata))
+
+    band = None
     raster = None
 
 
@@ -1182,8 +1348,8 @@ def zonal_statistics(
             aggregation coverage close to optimally by rasterizing sets of
             polygons that don't overlap.  However, this step can be
             computationally expensive for cases where there are many polygons.
-              this flag to False directs the function rasterize in one
-            step.
+            Setting this flag to ``False`` directs the function rasterize in
+            one step.
         include_value_counts (boolean): If True, the function tallies the
             number of pixels of each value under the polygon.  This is useful
             for classified rasters but could exhaust available memory when run
@@ -1648,6 +1814,10 @@ def get_raster_info(raster_path):
           efficient reading.
         * ``'numpy_type'`` (numpy type): this is the equivalent numpy datatype
           for the raster bands including signed bytes.
+        * ``'overviews'`` (sequence): A list of (x, y) tuples for the
+          number of pixels in the width and height of each overview level of
+          the raster.
+        * ``'file_list'`` (sequence): A list of files that make up this raster.
 
     """
     raster = gdal.OpenEx(raster_path, gdal.OF_RASTER)
@@ -1670,6 +1840,15 @@ def get_raster_info(raster_path):
     raster_properties['nodata'] = [
         raster.GetRasterBand(index).GetNoDataValue() for index in range(
             1, raster_properties['n_bands']+1)]
+
+    # GDAL creates overviews for the whole raster but has overviews accessed
+    # per band.  We assume that all bands have the same overviews.
+    raster_properties['overviews'] = []
+    for overview_index in range(raster.GetRasterBand(1).GetOverviewCount()):
+        overview_band = raster.GetRasterBand(1).GetOverview(overview_index)
+        raster_properties['overviews'].append((
+            overview_band.XSize, overview_band.YSize))
+
     # blocksize is the same for all bands, so we can just get the first
     raster_properties['block_size'] = raster.GetRasterBand(1).GetBlockSize()
 
@@ -1707,11 +1886,17 @@ def get_raster_info(raster_path):
 def reproject_vector(
         base_vector_path, target_projection_wkt, target_path, layer_id=0,
         driver_name='ESRI Shapefile', copy_fields=True,
+        target_layer_name=None,
         osr_axis_mapping_strategy=DEFAULT_OSR_AXIS_MAPPING_STRATEGY):
     """Reproject OGR DataSource (vector).
 
     Transforms the features of the base vector to the desired output
-    projection in a new ESRI Shapefile.
+    projection in a new vector.
+
+    Note:
+        If the ESRI Shapefile driver is used, the ``target_layer_name``
+        optional parameter is ignored. ESRI Shapefiles by definition use the
+        filename to define the layer name.
 
     Args:
         base_vector_path (string): Path to the base shapefile to transform.
@@ -1727,12 +1912,15 @@ def reproject_vector(
             reprojection step. If it is an iterable, it will contain the
             field names to exclusively copy. An unmatched fieldname will be
             ignored. If ``False`` no fields are copied into the new vector.
+        target_layer_name=None (str): The name to use for the target layer in
+            the new vector.  If ``None`` (the default), the layer name from the
+            source layer will be used.
         osr_axis_mapping_strategy (int): OSR axis mapping strategy for
             ``SpatialReference`` objects. Defaults to
             ``geoprocessing.DEFAULT_OSR_AXIS_MAPPING_STRATEGY``. This parameter
             should not be changed unless you know what you are doing.
 
-    Return:
+    Returns:
         None
     """
     base_vector = gdal.OpenEx(base_vector_path, gdal.OF_VECTOR)
@@ -1754,8 +1942,20 @@ def reproject_vector(
 
     # Create new layer for target_vector using same name and
     # geometry type from base vector but new projection
+    layer_name = layer_dfn.GetName()
+    if target_layer_name is not None:
+        layer_name = target_layer_name
+        if driver_name == 'ESRI Shapefile':
+            target_file_basename = os.path.splitext(
+                os.path.basename(target_path)[0])
+            if layer_name != target_file_basename:
+                LOGGER.warning(
+                    f'Ignoring user-defined layer name {layer_name}. '
+                    f'Defining a layer name is incompatible with the ESRI '
+                    'Shapefile vector format.  Use the filename instead or '
+                    'use a different vector format.')
     target_layer = target_vector.CreateLayer(
-        layer_dfn.GetName(), target_sr, layer_dfn.GetGeomType())
+        layer_name, target_sr, layer_dfn.GetGeomType())
 
     # this will map the target field index to the base index it came from
     # in case we don't need to copy all the fields
@@ -1953,7 +2153,7 @@ def warp_raster(
         gdal_warp_options=None, working_dir=None,
         raster_driver_creation_tuple=DEFAULT_GTIFF_CREATION_TUPLE_OPTIONS,
         osr_axis_mapping_strategy=DEFAULT_OSR_AXIS_MAPPING_STRATEGY):
-    """Resize/resample raster to desired pixel size, bbox and projection.
+    f"""Resize/resample raster to desired pixel size, bbox and projection.
 
     Args:
         base_raster_path (string): path to base raster.
@@ -1962,7 +2162,7 @@ def warp_raster(
         target_raster_path (string): the location of the resized and
             resampled raster.
         resample_method (string): the resampling technique, one of
-            ``near|bilinear|cubic|cubicspline|lanczos|average|mode|max|min|med|q1|q3``
+            ``{_GDAL_WARP_ALGOS_FOR_HUMAN_EYES}``
         target_bb (sequence): if None, target bounding box is the same as the
             source bounding box.  Otherwise it's a sequence of float
             describing target bounding box in target coordinate system as
@@ -2117,11 +2317,10 @@ def warp_raster(
             'PIXELTYPE' not in ' '.join(raster_creation_options)):
         raster_creation_options.append('PIXELTYPE=SIGNEDBYTE')
 
-    # WarpOptions.this is None when an invalid option is passed, and it's a
-    # truthy SWIG proxy object when it's given a valid resample arg.
-    if not gdal.WarpOptions(resampleAlg=resample_method)[0].this:
+    if resample_method.lower() not in _GDAL_WARP_ALGORITHMS:
         raise ValueError(
-            f'Invalid resample method: "{resample_method}"')
+            f'Invalid resample method: "{resample_method}". '
+            f'Must be one of {_GDAL_WARP_ALGOS_FOR_HUMAN_EYES}')
 
     gdal.Warp(
         warped_raster_path, base_raster,
@@ -2276,7 +2475,8 @@ def rasterize(
 
 
 def calculate_disjoint_polygon_set(
-        vector_path, layer_id=0, bounding_box=None):
+        vector_path, layer_id=0, bounding_box=None,
+        geometries_may_touch=False):
     """Create a sequence of sets of polygons that don't overlap.
 
     Determining the minimal number of those sets is an np-complete problem so
@@ -2291,6 +2491,10 @@ def calculate_disjoint_polygon_set(
             does not intersect this bounding box it will not be considered
             in the disjoint calculation. Coordinates are in the order
             [minx, miny, maxx, maxy].
+        geometries_may_touch=False(bool): If ``True``, geometries in a subset
+            are allowed to have touching boundaries, but are not allowed to
+            have intersecting interiors.  If ``False`` (the default), no
+            geometries in a subset may intersect in any way.
 
     Return:
         subset_list (sequence): sequence of sets of FIDs from vector_path
@@ -2365,6 +2569,12 @@ def calculate_disjoint_polygon_set(
         else:
             polygon = poly_geom
         for intersect_poly_fid in possible_intersection_set:
+            # If geometries touch (share 1+ boundary point), then do not count
+            # it as an intersection.
+            if geometries_may_touch and polygon.touches(
+                    shapely_polygon_lookup[intersect_poly_fid]):
+                continue
+
             if intersect_poly_fid == poly_fid or polygon.intersects(
                     shapely_polygon_lookup[intersect_poly_fid]):
                 poly_intersect_lookup[poly_fid].add(intersect_poly_fid)
@@ -2684,10 +2894,12 @@ def convolve_2d(
     signal_raster_info = get_raster_info(signal_path_band[0])
     kernel_raster_info = get_raster_info(kernel_path_band[0])
 
-    for info_dict in [signal_raster_info, kernel_raster_info]:
+    for info_dict, raster_path_band in zip(
+            [signal_raster_info, kernel_raster_info],
+            [signal_path_band, kernel_path_band]):
         if 1 in info_dict['block_size']:
             raise ValueError(
-                f'{signal_path_band} has a row blocksize which can make this '
+                f'{raster_path_band} has a row blocksize which can make this '
                 f'function run very slow, create a square blocksize using '
                 f'`warp_raster` or `align_and_resize_raster_stack` which '
                 f'creates square blocksizes by default')
@@ -3321,15 +3533,19 @@ def get_gis_type(path):
     Args:
         path (str): path to a file on disk.
 
+    Raises:
+        ValueError
+            if ``path`` is not a file or cannot be opened as a
+            ``gdal.OF_RASTER`` or ``gdal.OF_VECTOR``.
 
     Return:
         A bitwise OR of all GIS types that PyGeoprocessing models, currently
-        this is ``pygeoprocessing.UNKNOWN_TYPE``,
+        this is
         ``pygeoprocessing.RASTER_TYPE``, or ``pygeoprocessing.VECTOR_TYPE``.
 
     """
     if not os.path.exists(path):
-        raise ValueError("%s does not exist", path)
+        raise ValueError(f"{path} does not exist.")
     from pygeoprocessing import UNKNOWN_TYPE
     gis_type = UNKNOWN_TYPE
     gis_raster = gdal.OpenEx(path, gdal.OF_RASTER)
@@ -3341,6 +3557,10 @@ def get_gis_type(path):
     if gis_vector is not None:
         from pygeoprocessing import VECTOR_TYPE
         gis_type |= VECTOR_TYPE
+        gis_vector = None
+    if gis_type == UNKNOWN_TYPE:
+        raise ValueError(
+            f"Could not open {path} as a gdal.OF_RASTER or gdal.OF_VECTOR.")
     return gis_type
 
 
@@ -3749,7 +3969,7 @@ def stitch_rasters(
         overlap_algorithm='etch',
         area_weight_m2_to_wgs84=False,
         osr_axis_mapping_strategy=DEFAULT_OSR_AXIS_MAPPING_STRATEGY):
-    """Stitch the raster in the base list into the existing target.
+    f"""Stitch the raster in the base list into the existing target.
 
     Args:
         base_raster_path_band_list (sequence): sequence of raster path/band
@@ -3757,7 +3977,7 @@ def stitch_rasters(
         resample_method_list (sequence): a sequence of resampling methods
             which one to one map each path in ``base_raster_path_band_list``
             during resizing.  Each element must be one of
-            "near|bilinear|cubic|cubicspline|lanczos|mode".
+            ``{_GDAL_WARP_ALGOS_FOR_HUMAN_EYES}``
         target_stitch_raster_path_band (tuple): raster path/band tuple to an
             existing raster, values in ``base_raster_path_band_list`` will
             be stitched into this raster/band in the order they are in the
@@ -4041,6 +4261,107 @@ def stitch_rasters(
 
     target_raster = None
     target_band = None
+
+
+def build_overviews(
+        raster_path, internal=False, resample_method='near',
+        overwrite=False, levels='auto'):
+    f"""Build overviews for a raster dataset.
+
+    Args:
+        raster_path (str): A path to a raster on disk for which overviews
+            should be built.
+        internal=False (bool): Whether to modify the raster when building
+            overviews. In GeoTiffs, this builds internal overviews when
+            ``internal=True``, and external overviews when ``internal=False``.
+        resample_method='near' (str): The resample method to use when
+            building overviews.  Must be one of
+            ``{_GDAL_WARP_ALGOS_FOR_HUMAN_EYES}``.
+        overwrite=False (bool): Whether to overwrite existing overviews, if
+            any exist.
+        levels='auto' (sequence): A sequence of integer overview levels. If
+            ``'auto'``, overview levels will be determined by using factors of
+            2 until the overview's x and y dimensions are both less than 256.
+
+    Example:
+        Generate overviews, regardless of whether overviews already exist
+        for the raster, letting the function determine the levels of overviews
+        to generate::
+
+            build_overviews(raster_path)
+
+        Generate overviews for 4 levels, at 1/2, 1/4, 1/8 and 1/16 the
+        resolution::
+
+            build_overviews(raster_path, levels=[2, 4, 8, 16])
+
+    Returns:
+        ``None``
+    """
+    if resample_method.lower() not in _GDAL_WARP_ALGORITHMS:
+        raise ValueError(
+            f'Invalid overview resample method: "{resample_method}". '
+            f'Must be one of {_GDAL_WARP_ALGOS_FOR_HUMAN_EYES}')
+
+    def overviews_progress(*args, **kwargs):
+        pct_complete, name, other = args
+        percent = round(pct_complete * 100, 2)
+        if time.time() - overviews_progress.last_progress_report > 5.0:
+            LOGGER.info(f"Overviews progress: {percent}%")
+            overviews_progress.last_progress_report = time.time()
+    overviews_progress.last_progress_report = time.time()
+
+    open_flags = gdal.OF_RASTER
+    if internal:
+        open_flags |= gdal.GA_Update
+        LOGGER.info(f"Building internal overviews on {raster_path}")
+    else:
+        LOGGER.info("Building external overviews.")
+    raster = gdal.OpenEx(raster_path, open_flags)
+    overview_count = 0
+    for band_index in range(1, raster.RasterCount + 1):
+        band = raster.GetRasterBand(band_index)
+        overview_count += band.GetOverviewCount()
+
+    if overview_count > 0:
+        if overwrite:
+            LOGGER.info(f"Clearing existing overviews from {raster_path}")
+            result = raster.BuildOverviews(
+                resampling=resample_method,
+                overviewlist=[],
+                callback=overviews_progress
+            )
+            LOGGER.info(f"Overviews cleared from {raster_path}")
+        else:
+            raise ValueError(
+                f"Raster already has overviews.  Use "
+                "overwrite=True to override this and regenerate overviews on "
+                f"{raster_path}")
+
+    # This loop and limiting factor borrowed from gdaladdo.cpp.
+    # Create overviews so long as the overviews are at least 256 pixels in
+    # either x or y dimensions.
+    if levels == 'auto':
+        overview_scales = []
+        factor = 2
+        limiting_factor = 256
+        while (math.ceil(raster.RasterXSize / factor) > limiting_factor or
+               math.ceil(raster.RasterYSize / factor) > limiting_factor):
+            overview_scales.append(factor)
+            factor *= 2
+    else:
+        overview_scales = [int(level) for level in levels]
+
+    LOGGER.debug(f"Using overviews {overview_scales}")
+    result = raster.BuildOverviews(
+        resampling=resample_method,
+        overviewlist=overview_scales,
+        callback=overviews_progress
+    )
+    LOGGER.info(f"Overviews completed for {raster_path}")
+    if result:  # Result will be nonzero on error.
+        raise RuntimeError(
+            f"Building overviews failed or was interrupted for {raster_path}")
 
 
 def _m2_area_of_wg84_pixel(pixel_size, center_lat):